Tablet verison fixed. App mostly finished
The tablet layout is functional, it uses several fragments to display
content in one screen. This required the removal of
activity_itunesitem_twopane and the addition of
fragment_slider_pager_twopane. Changed the share button's content.
Additional ref files added to indicate when the device is a tablet or
smartphone (or somewhere in between).