ItunesItems loaded from Json are now stored in CategoryObjects
Instead of storing JSON responses in Bundles that are reformed every
time the fragment restarts, responses are now stored in category objects
which are globally accessible form the singleton ItunesAppController.
The result is faster loading of list views and it seems to have fixed
the glitch where items do not load on the first bootup of the app.