Reformatted Intent communication
Cut down on passing serialized items via intents because they were
slowing down the application. Therefore, CategoryAtttribute is no longer
Serializable. UserFavorites are now stored in an ArrayList instead of a
Map, which makes it easier to manipulate the items when necessary. Entry
now impliments a custom equals method. The adview is disabled until I
can figure out why it crashes on tablets.